requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
Flask-Cors
pandas
SQLAlchemy
openpyxl
numpy
matplotlib
requests
pytest
python-dotenv
pydantic
psycopg2-binary
pytest
pytest-mock